fix: mock data to separate file
